refactor code for compositions docs examples
