+++ scopes/compositions/compositions/docs/example-layout/index.ts
@@ -1,0 +1,1 @@
+export { ExampleLayout } from './example-layout';
