+++ scopes/compositions/compositions/docs/user-hero/index.ts
@@ -1,0 +1,1 @@
+export { UserHero, UserHeroExample } from './user-hero';
