+++ scopes/compositions/compositions/docs/user-hero/user-hero-string.ts
@@ -1,0 +1,18 @@
+export const userHeroString = `
+// user-hero.compositions.jsx
+
+import { Hero } from './hero';
+
+const UserHeroWithDescription = () => {
+  return (
+    <Hero
+      title="Jessica Pegula"
+      description="Frontend developer and designer."
+      profileImage={img}
+      data-testid="test-hero"
+      userName="jessica"
+    />
+  );
+};
+
+`;
