+++ scopes/compositions/compositions/docs/user-profile/index.ts
@@ -1,0 +1,1 @@
+export { UserProfileExample } from './user-profile';
