+++ scopes/compositions/compositions/docs/user-profile/scope-data.ts
@@ -1,0 +1,26 @@
+export const scopeData = [
+  {
+    id: 'teambit.base-ui',
+    description: 'A collection of minimal, bare, and well-crafted UI elements.',
+    componentCount: '68',
+    visibility: 'Public',
+  },
+  {
+    id: 'teambit.base-ui',
+    description: 'A collection of minimal, bare, and well-crafted UI elements.',
+    componentCount: '68',
+    visibility: 'Public',
+  },
+  {
+    id: 'teambit.evangelist',
+    description: 'Reusable UI components for marketing projects.',
+    componentCount: '54',
+    visibility: 'Public',
+  },
+  {
+    id: 'teambit.evangelist',
+    description: 'Reusable UI components for marketing projects.',
+    componentCount: '54',
+    visibility: 'Public',
+  },
+];
